unique key의 경우 modelform을 이용하면 자꾸 form_invalid를 타고 있었음.(동일한 field value를 제출한 경우) 우선 email은 form에서 빼고, unsername은 unique조건을 빼버림.
photo의 경우 사용자가 변경하지 않았으면 user table에서도 변경하지 않는 로직 추가(다만 empty value제출을 위해 required=false를 추가)

+++ app/views.py
@@ -1,4 +1,5 @@
 from django.views.generic import FormView, TemplateView
+from django.forms import ValidationError
 from .forms import PhotoUploadForm
 from django.shortcuts import redirect
 import os
@@ -13,14 +14,16 @@ class EditView(FormView):
     def form_valid(self, form):
         user = self.request.user
 
-        if os.path.isfile(user.photo.path):
-            os.remove(user.photo.path)
-
-        user.photo = form.cleaned_data['photo']
+        if 'photo' in form.changed_data:
+            if os.path.isfile(user.photo.path):
+                os.remove(user.photo.path)
+            user.photo = form.cleaned_data['photo']
+        
         user.username = form.cleaned_data['username']
-        user.email = form.cleaned_data['email']
         user.save()
-
+    
         return redirect('app:profile')
 
+    def form_invalid(self, form):
+        raise ValidationError(form.errors)
